requirements.txt
pytest
pytest-timeout
beautifulsoup4
mysql-connector-python
spacy
nltk
flask
flask-cors
flask-sqlalchemy
pymysql